refactor: BoardPage 周辺 hook の責務整理
- useBoardPageResources・useBoardPageInteractions を削除し useBoardPageController へ直接インライン化
- handleAdded・handleWorksAdded（同一ロジック）を handleNavigateToStacked へ統合
