+++ src/features/backlog/hooks/useBoardPageController.ts
@@ -1,7 +1,9 @@
 import type { Session } from "@supabase/supabase-js";
-import { useBoardPageInteractions } from "./useBoardPageInteractions.ts";
+import { useBacklogActions } from "./useBacklogActions.ts";
+import { useBacklogDnd } from "./useBacklogDnd.ts";
+import { useBacklogFeedback } from "./useBacklogFeedback.tsx";
+import { useBacklogItems } from "./useBacklogItems.ts";
 import { useBoardPageLayout } from "./useBoardPageLayout.ts";
-import { useBoardPageResources } from "./useBoardPageResources.ts";
 import { useBoardPageState } from "./useBoardPageState.ts";
 
 type UseBoardPageControllerOptions = {
@@ -10,20 +12,23 @@ type UseBoardPageControllerOptions = {
 
 export function useBoardPageController({ session }: UseBoardPageControllerOptions) {
   const { isMobileLayout } = useBoardPageLayout();
-  const { feedback, feedbackUi, items, setItems, isLoading, error, loadItems } =
-    useBoardPageResources();
-  const boardPageState = useBoardPageState({
+  const { feedback, feedbackUi } = useBacklogFeedback();
+  const { items, setItems, isLoading, error, loadItems } = useBacklogItems();
+  const boardPageState = useBoardPageState({ isMobileLayout, setItems });
+
+  const dnd = useBacklogDnd({
+    items,
     isMobileLayout,
-    setItems,
+    onAfterDrop: loadItems,
+    feedback,
   });
 
-  const { dnd, actions } = useBoardPageInteractions({
+  const actions = useBacklogActions({
     items,
     session,
-    isMobileLayout,
     loadItems,
     onItemDeleted: boardPageState.handleItemDeleted,
-    onWorksAdded: boardPageState.handleWorksAdded,
+    onWorksAdded: boardPageState.handleNavigateToStacked,
     feedback,
   });
 
@@ -64,7 +69,7 @@ export function useBoardPageController({ session }: UseBoardPageControllerOption
       session,
       onClose: boardPageState.handleCloseAddModal,
       onAdded: () => {
-        boardPageState.handleAdded();
+        boardPageState.handleNavigateToStacked();
         void loadItems();
       },
     },
